testing priority scheduler
